pyproject.toml
[project]
name = "library-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <3.14"
dependencies = [
    "alembic>=1.18.4",
    "flask>=3.1.3",
    "flask-injector>=0.15.0",
    "injector>=0.24.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
]
